ivr-backend/requirements.txt
sounddevice
numpy
openai-whisper
soundfile
noisereduce
scipy
scipy
torch
rasa
requests
python-decouple
uvicorn
fastapi
pydub
ffmpeg